requirements.txt
python-dotenv
openai
pydantic
fastapi
uvicorn
rapidfuzz
lingua-language-detector
dateparser
apscheduler
loguru
sentence-transformers
numpy
pytest
pytest-asyncio
edge-tts
faster-whisper
python-multipart
